requirements.txt
discord.py>=1.6.0
audioread>=2.1.9